data-import/requirements.txt
xlrd==2.0.1
psycopg2-binary==2.9.12
python-dotenv==1.0.0
tqdm==4.66.1
pandas>=2.0.0
odfpy>=1.4.0
